Separate logging classes into separate files
