+++ csharp/extractor/Semmle.Util/Logging/Severity.cs
@@ -1,0 +1,14 @@
+namespace Semmle.Util.Logging
+{
+    /// <summary>
+    /// The severity of a log message.
+    /// </summary>
+    public enum Severity
+    {
+        Trace = 1,
+        Debug = 2,
+        Info = 3,
+        Warning = 4,
+        Error = 5
+    }
+}
