+++ csharp/extractor/Semmle.Util/Logging/Verbosity.cs
@@ -1,0 +1,16 @@
+namespace Semmle.Util.Logging
+{
+    /// <summary>
+    /// Log verbosity level.
+    /// </summary>
+    public enum Verbosity
+    {
+        Off = 0,
+        Error = 1,
+        Warning = 2,
+        Info = 3,
+        Debug = 4,
+        Trace = 5,
+        All = 6
+    }
+}
